chore: rename folders to better reflect what it is
